tests/tests/Clipperz/Crypto/RSA.html

<!--

Copyright 2008-2014 Clipperz Srl - All rights reserved.

This file is part of Clipperz's Javascript Crypto Library.

The Javascript Crypto Library provides web developers with an extensive 
and efficient set of cryptographic functions.
For further information about its features and functionalities please
refer to https://clipperz.is

* The Javascript Crypto Library is free software: you can redistribute
  it and/or modify it under the terms of the BSD 3-Clause License
  as published by the Free Software Foundation.

* You should have received a copy of the BSD 3-Clause License
  along with the Javascript Crypto Library in the LICENSE.TXT file. 
  If not, see <http://www.gnu.org/licenses/>.

-->

<html>
<head>
	<script type="text/javascript" src="../../../../other_libraries/MochiKit/MochiKit.js"></script>
    <script type="text/javascript" src="../../../SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="../../../SimpleTest/test.css">

	<script type='text/javascript' src='../../../../js/Clipperz/Base.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/DOM.js'></script>

	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/BigInt.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/Base.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/RSA.js'></script>

</head>
<body>
<pre id="test">
<script type="text/javascript">
try {
	var	rsaPublicKey;
	var	randomValue;
	var	publicEncryptedValue;
	var privateEncryptedValue;

	var	startTime;
	var	endTime;


startTime=new Date();
	rsaPublicKey = Clipperz.Crypto.RSA.generatePublicKey(512);

endTime=new Date();
is(true, true, 'Time to generate the RSA public key (size 512 bit): '+(endTime.getTime()-startTime.getTime())/1000.0);

	randomValue = Clipperz.Crypto.Base.generateRandomSeed();
	publicEncryptedValue = Clipperz.Crypto.RSA.encryptUsingPublicKey(rsaPublicKey, randomValue);
	privateEncryptedValue = Clipperz.Crypto.RSA.encryptUsingPrivateKey(rsaPublicKey, randomValue);
	
	is(publicEncryptedValue == privateEncryptedValue, false);
	is(Clipperz.Crypto.RSA.decryptUsingPrivateKey(rsaPublicKey, publicEncryptedValue), randomValue);
	is(Clipperz.Crypto.RSA.decryptUsingPublicKey(rsaPublicKey, privateEncryptedValue), randomValue);

	//-------------------------------------------------------------------------
} catch (err) {
    
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );
}

</script>
</pre>
</body>
</html>
